% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCrawlerTW711.R
\name{getCityCodes}
\alias{getCityCodes}
\title{getCityCodes via POST cityid}
\usage{
getCityCodes()
}
\description{
getCityCodes via POST cityid
}
\examples{
cityCodeDf = getCityCodes()
View(cityCodeDf)
}

