% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCrawlerTW711.R
\name{getStoreData}
\alias{getStoreData}
\title{getStoreData via POST cityName and townName}
\usage{
getStoreData(cityName, townName)
}
\description{
getStoreData via POST cityName and townName
}
\examples{
cityCodeDf = getCityCodes()
townDf = getTownDf(cityCodeDf$cityCode)
testTownDf = townDf[,c("cityName","TownName")][1:10,]
listOfData = apply(testTownDf,1,function(x){
   print(getStoreData(x[1],x[2]))
})
totalDf = do.call(rbind,listOfData)
View(totalDf)
}

