% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RCrawlerTW711.R
\name{getTownDf}
\alias{getTownDf}
\title{getTownDf via POST cityCodes}
\usage{
getTownDf(cityCodes)
}
\description{
getTownDf via POST cityCodes
}
\examples{
cityCodeDf = getCityCodes()
townDf = getTownDf(cityCodeDf$cityCode)
View(townDf)
}

